PYTRAJ: fix memory stuff
Reason:
if we keep calling `calculate`, the data will be appended to a hidden DataSetList
object and return the wrong size. I just did a trick, like
"return dslist[old_size:new_size]". This will return a memoryview for a
chunk of newly updated datasets.

Not very good solution but need to solve to throw it out of my mind. :D