PYTRAJ: add indice slicing for Topology
We have topology['@CA'] that returns a list of CA atom objects.

we can also use topology[indice] where `indice` is an array/list/tuple ...

this returns a list of atoms having index in `indices`

I add this so pytraj can "steal" indices from other's packages and feed to cpptraj
to do analysis.