PYTRAJ: fix search_hbonds. need to call "Print"
From Dan:
"""I think for some reason then the Print()
routine isn't being called from pytraj. The issue here is that hbonds
can only be updated in the DoAction() routine when they are detected.
So if an hbond is formed in the beginning of a trajectory but is not
formed for the last few frames you could end up with the array coming
up short. So in Print() there is a section that brings any such hbonds
up to speed (see lines 732-745 in Action_Hbond.cpp)"""

And I just updated "print_output" in hbonds.py. Worked.

now you can just "traj.search_hbonds().to_dict()" (or to_list(), to_ndarray())