pytraj: split_iterators for TrajectoryIterator
it's crazy, but I think this is cool. TrajectoryIterator is a generator
and we still can split it into different chunk (with autoimage, rmsfit).

This is different from chunk_iter, which load a chunk of frames into memory.
`split_iterators` create n_chunks of FrameIter objects. Only single frame will be
loaded at a time. Overall aim is to easily split TrajectoryIterator to different cores
for multiprocessing, mpi or whatever.